feat: Generalized DPI scaling
This works by defining closures for the key widgets and spacings, which are then hooked to a `dpi_change` signal. Computation no longer directly occurs in the DPI change handler

Old code has been commented. Signal handlers in `KeyboardWidget.__setup_ui` could be replaced with just the emit call to forward the signal to dpi_change, but it is unknown whether automatic window resizing is still desired behavior